add support for 0 length BC

+++ transposon/submit.sh
@@ -261,8 +261,10 @@ if [ ${runMerge} -eq 1 ]; then
     cat <<EOF | qsub -S /bin/bash -j y -b y -N merge.${sample} -o ${OUTDIR} -terse ${mergeHold} | perl -pe 's/[^\d].+$//g;' > sgeid.merge.${sample}
     set -eu -o pipefail
     echo "Merging barcodes"
-    zcat -f ${bcfiles} | pigz -p ${NSLOTS} -c -9 > $OUTDIR/${sample}.barcodes.preFilter.txt.gz
-    rm -f ${bcfiles}
+    if [[ ${bclen} -gt 0 ]]; then
+        zcat -f ${bcfiles} | pigz -p ${NSLOTS} -c -9 > $OUTDIR/${sample}.barcodes.preFilter.txt.gz
+        rm -f ${bcfiles}
+    fi
     
     if [[ ${sampleType} == "iPCR" ]]; then
         echo "Merging bam files"
@@ -298,10 +300,13 @@ fi
 #-o ${OUTDIR} breaks Jesper's Flowcell_Info.sh
 cat <<EOF | qsub -S /bin/bash -j y -b y -N ${sample} -terse ${analysisHold} # | perl -pe 's/[^\d].+$//g;' > sgeid.analysis
 set -eu -o pipefail
-${src}/analyzeBCcounts.sh ${minReadCutoff} ${sample}
 
-if [[ ${sampleType} == "iPCR" ]]; then
-    ${src}/analyzeIntegrations.sh ${sample}
+if [[ ${bclen} -gt 0 ]]; then
+    ${src}/analyzeBCcounts.sh ${minReadCutoff} ${sample}
+
+    if [[ ${sampleType} == "iPCR" ]]; then
+        ${src}/analyzeIntegrations.sh ${sample}
+    fi
 fi
 EOF
 rm -f sgeid.merge.${sample}
